requirements.txt
speedtest-cli==2.1.3
tomli>=2.0.1; python_version < "3.11"

